Adding Resource Management to R bridge
R code executed on CHTC seems to automatically think it is entitled to the entire server's ram. In order to account for this, the R bridge injects resource management code to the start of every R script, limiting the maximum memory. This also expands the max size of future.globals for more intense processes. Both of these limits are set to the value of r_memory_limit_gb specified under pipeline in the config.

In order to integrate and globally scope this, the config file was modified to append a reasonable default of 8 if nothing is specified (this default is used elsewhere as well for error handling). The executor class has been modified to write this value to the data context under the label "R_MEMORY_LIMIT_GB". Capitals are used here to indicate setting information rather than data being used/generated by the modules. Modules written to use the R bridge can then retrieve this setting from the data context when creating an R bridge. Modules can also manually specify memory limits, although perhaps this behavior should be used in tandem with checking the existing limit in order to avoid exceeding it.

Changes were made to soupX module flow to continue testing these. This component is still not functional.