Beginning Report Generation
Beginning to lay the ground work of report generation and exporting. Created a structure to handle this in the data context. Added a method `add_figure()`, which appends a figure with a selection of attributes to be displayed in the markdown document, formatted accordingly. REPORT_FIGURES is a list that contains a list of each module that generates a figure.