Paramater Schemas and validation, Run pipeline temp fix
Adding parameter schemas to top of each module and validation as a method of base class module.

Honestly this probably should go somewhere else to make it easier to validate parameters at the top of runtime, but that feels like an easy refactor that can get kicked down the can later. Worst case scenario is it takes longer to in the run to find out your config had a typo I guess.

Also added some makeshift code to make the run script work. Eventually I'll probably wrap this whole thing as a package and install it, which is how I think this type of thing is supposed to work, but Idrk how that works yet and it feels like a later problem.