Implementing Filtering Module - largely untested as of yet
Adding the filtering module. This has a lot of configurable pieces that still need to be tested at large. There's just a lot of code here...

The filtering module takes the a dictionary of filters with names as keys in its parameters that fall under the three types for now: Numeric, Bool, or Categorical. Additional preferences are under each individual metric being filtered. The metric name is assumed to be the same as the key of the dictionary unless otherwise specified as a 'column' field in that particular entry. For numeric types, mins and maxes can be specified as well as automatic outlier detection. Outlier detection uses mean absolute deviations to identify putative outliers, which can then be optionally removed.

There is a create_plots function under this module as well, but I suspect the formatting for this will need improvement.