Deleting Sphinx docs dependencies and refactoring the way layers are used
Giving up on Sphinx docs generation

Also reworking how adata layers are used. Before this change, layers were not saved when changes to the active matrix were made, requiring later modules that created to layers to try and save the active layer without knowing what that layer contained. This commit introduces adata_utils to make it easier to save and switch between layers while also adding some reference of the name of the active layer ( which works as long as someone doesn't manually change active layer in their own module).

I am now in the progress of adding layer_key as a parameter to most modules that might want to operate on a specific layer of the adata.