Adding clustering to dim reduction and coloring
Not entirely sure if it makes sense to have clustering be a part of this module, but it sort of needs to happen after neighbors and before plotting umap and tsne?

Alternative to this approach is to remove clustering from this and add it as a separate module. Then, plotting of dim reductions would be a standalone module that occurs after? This might make more functional sense in the long run, but would require more effort right now...